feat(Logger): Allow for dynamic log levels to be resolved by TypeScript compiler
On initialization, the Logger instance will define methods for handling each of log levels by "filling in" the first parameter of the `log()` method with the level. When using TypeScript in a dependent project, those methods cannot be resolved. This feature allows resolution, at the cost of typings. More research could be done to try and get around this limitation at a later date.